requirements.txt
discord.py
colorizelib